(Bugfix) Handle Background Patches in Instanseg Inference
When processing bounding boxes with no cells, the error occurs in the
centroids_from_lab function when sparse_onehot is empty. The problem is in the
line: sum_centroids = torch.sparse.mm(sparse_onehot, mesh_grid.flatten(1).T) In
the get_masked_patches function, when there are no cells in the bounding box,
the tensor lab has a max value of 0, which causes issues with the tensor
dimensions in the sparse matrix multiplication. conditional block to handle
empty patches